Sizzle: warn on selector de-optimization
When this code path is reached, a setAttribute("id") is called followed by removing that same attribute. Since we are mutating the DOM, this will cause a style recalculation, which can be very expensive in certain scenarios. It may also be surprising to users, since they might not expect an element.find() to cause DOM mutations. This PR adds a console warning when the de-optimized code path is reached.