Build: Update most dependencies
Karma can't be updated to 1.4 as it breaks on IE 8 & lower:
https://github.com/karma-runner/karma/issues/2556

Grunt can't be updated to 1.x as the old grunt-jscs peer-depends on Grunt 0.4
and in npm 2 (used by default in Node 4 LTS) this breaks npm install. Updating
grunt-jscs to a newer version (or switching to ESLint) would require lots
of style fixes.