Tests: Test for Shadow DOM v1 support
There was a test for Shadow DOM nodes working as root but it was using Shadow
DOM v0 that only Chromium implements and which it plans to remove in the future.
The test was reworked to rely on Shadow DOM v1.

jQuery master already uses the new API in its tests.