Core: Make Sizzle.contains work within <template/> elements
The `<template/>` element `contents` property is a document fragment that may
have a `null` `documentElement`. In Safari 16 this happens in more cases due
to recent spec changes - in particular, even if that document fragment is
explicitly adopted into an outer document. We're testing both of those cases
now.

The above behavior made `Sizzle.contains` crash when run on certain elements
within the `<template/>` element. As it turns out, we don't need
to query the supposed container `documentElement` if it has the
`Node.DOCUMENT_NODE` (9) `nodeType`; we can call `.contains()` directly on
the `document`. That avoids the crash.

However, we still need to fall back to `documentElement` if one is
defined as IE <9 have a broken `.contains()` on the document.

Fixes jquery/jquery#5147
Closes gh-490
Ref jquery/jquery#5158